+++ designer/src/api/toolbox-api.ts
@@ -1,11 +1,12 @@
 import { Step } from '../definition';
 import { BehaviorController } from '../behaviors/behavior-controller';
 import { ObjectCloner, Uid, Vector } from '../core';
-import { StepDefinition, StepsConfiguration } from '../designer-configuration';
+import { StepDefinition } from '../designer-configuration';
 import { DesignerState } from '../designer-state';
 import { DragStepBehavior } from '../behaviors/drag-step-behavior';
 import { DesignerContext } from '../designer-context';
 import { LayoutController } from '../layout-controller';
+import { IconProvider } from '../core/icon-provider';
 
 export class ToolboxApi {
 	public constructor(
@@ -13,7 +14,7 @@ export class ToolboxApi {
 		private readonly designerContext: DesignerContext,
 		private readonly behaviorController: BehaviorController,
 		private readonly layoutController: LayoutController,
-		private readonly configuration: StepsConfiguration
+		private readonly iconProvider: IconProvider
 	) {}
 
 	public isVisibleAtStart(): boolean {
@@ -21,7 +22,7 @@ export class ToolboxApi {
 	}
 
 	public tryGetIconUrl(step: StepDefinition): string | null {
-		return this.configuration.iconUrlProvider ? this.configuration.iconUrlProvider(step.componentType, step.type) : null;
+		return this.iconProvider.getIconUrl(step);
 	}
 
 	/**